requirements.txt
selenium>=4.0.0
pandas>=2.0.0
img2pdf>=0.4.0
Pillow>=10.0.0
PyPDF2>=3.0.0
chromedriver-autoinstaller>=0.6.0
python-dotenv>=1.0.0 